requirements/main.txt
flask
flask-imp==5.3.6
flask-sqlalchemy
sqlalchemy
python-dotenv
pytz
toml
click
huey
gunicorn
supervisor
email-validator
mail-parser
requests
pyhead
validators
beautifulsoup4
vite-transporter